requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==3.2.9
djangorestframework==3.12.4
pickle-mixin==1.0.2
pytz==2021.3
redis==4.0.2
sqlparse==0.4.2
wrapt==1.13.3

ta~=0.9.0
matplotlib~=3.5.1
pandas~=1.2.0